refactor: 将字符画调取模块置于插件目录下
